More EF stuff

+++ Sadie.Game.Players/Friendships/PlayerFriendship.cs
@@ -1,13 +1,13 @@
 ﻿namespace Sadie.Game.Players.Friendships;
 
 public class PlayerFriendship(
-    int id,
+    long id,
     int originId,
     int targetId,
     PlayerFriendshipStatus status,
     PlayerFriendshipData targetData)
 {
-    public int Id { get; } = id;
+    public long Id { get; } = id;
     public PlayerFriendshipStatus Status { get; set; } = status;
     public int OriginId { get; } = originId;
     public int TargetId { get; } = targetId;
